Struct pseudo_bank::model::account::Account

pub struct Account {
    id: ClientId,
    balance: Decimal,
    is_locked: bool,
    disputes: Vec<Transaction>,
}

Holds the data needed for a client's account

Instead of having a separate field for held funds, this model has only a total balance and a list of current disputes. The available balance is determined by subtracting the disputed transactions from the current total balance. This provides a layer of safety over just using calculations on a couple of numeric fields, namely that a dispute could be resolved multiple times, or a transaction disputed multiple times.

With this method it should be impossible to accidentally dispute twice or resolve twice.

Note: There is currently no protection against deposits or withdrawals with the same transaction id. This would need to be implemented in the Store.

Performance Considerations

Calculating the available balance is fast as long as there aren't many disputes. This should be relatively rare, so the safety guarantees should be ultimately worth it.

Fields
id: ClientIdbalance: Decimalis_locked: booldisputes: Vec<Transaction>
Implementations

impl Account

pub fn new(id: ClientId) -> Account

pub fn get_available(&self) -> Decimal

total - disputed

pub fn get_held(&self) -> Decimal

Sum of all disputed transactions

pub fn is_locked(&self) -> bool

Whether or not new transactions can be applied to this account

pub fn get_id(&self) -> ClientId

pub fn apply_transaction(
&mut self,
transaction_store: &HashMap<(ClientId, TransactionId), Transaction>,
transaction: &Transaction
) -> Result<bool, TransactionError>

Alters the current balances using a given transaction.

This is where the bulk of the processing in the application is done. A clients account can be credited, depited, disputed, and locked through this method.

Returns whether or not the transaction should be stored to the transaction list.

let transaction = Transaction {
    transaction_type: TransactionType::Deposit,
    client: client_id,
    tx: transaction_id,
    amount: Some(10_000.into()),
};

account.apply_transaction(transaction_store, transaction);

assert_eq!(account.get_available(), Decimal::from(10_000));